builder: split builder.go into 3 files, commands and command_support are for dockerfile commands
Docker-DCO-1.1-Signed-off-by: Erik Hollensbe <github@hollensbe.org> (github: erikh)
